+++ app/src/gui/home-tab/home-tab.builder.ts
@@ -1,28 +1,37 @@
 import { Injectable } from "@nestjs/common";
 import { BlockAction } from "@slack/bolt";
-import { Header, HomeTab, ViewBlockBuilder } from "slack-block-builder";
+import { Actions, Button, Header, HomeTab, ViewBlockBuilder } from "slack-block-builder";
 import { Appendable, SlackHomeTabDto } from "slack-block-builder/dist/internal";
-import { AppHomeOpenedArgs } from "../../../bolt/types/app-home-opened.type";
-import { ViewsPublishResponse } from "../../../bolt/types/views-publish-response.type";
-import { WebClient } from "../../../bolt/types/web-client.type";
-import { BlockBuilder } from "../../block-builder.interface";
-import { DevUiBuilder } from "../../dev/dev-ui.builder";
-import { DayListBuilder } from "./day-list.builder";
-import { VisibleOfficeSelectBuilder } from "./visible-office-select.builder";
+import { AppHomeOpenedArgs } from "../../bolt/types/app-home-opened.type";
+import { ViewsPublishResponse } from "../../bolt/types/views-publish-response.type";
+import { WebClient } from "../../bolt/types/web-client.type";
+import { DevUiBuilder } from "../dev/dev-ui.builder";
+import { DayList } from "./views/day-list.builder";
+import { VisibleOfficeSelect } from "./views/visible-office-select.builder";
 
 @Injectable()
-export class HomeTabBuilder implements BlockBuilder<ViewBlockBuilder> {
+export class HomeTabBuilder {
   constructor(
-    private dayListBlocks: DayListBuilder,
+    private dayListBlocks: DayList,
     private devToolsBuilder: DevUiBuilder,
-    private visibleOfficeSelectBuilder: VisibleOfficeSelectBuilder,
+    private visibleOfficeSelectBuilder: VisibleOfficeSelect,
   ) {}
 
   async buildBlocks(userId: string): Promise<Appendable<ViewBlockBuilder>> {
     const devTools = this.devToolsBuilder.buildBlocks();
     const officeSelect = await this.visibleOfficeSelectBuilder.buildBlocks();
     const dayList = await this.dayListBlocks.buildBlocks(userId);
-    return [...devTools, Header({ text: "Ilmoittautumiset" }), ...officeSelect, ...dayList];
+    return [
+      ...devTools,
+      Actions().elements([
+        Button({ text: "Ilmoittautuminen" }),
+        Button({ text: "Läsnäolijat" }),
+        Button({ text: "Asetukset" }),
+      ]),
+      Header({ text: "Ilmoittautuminen" }),
+      ...officeSelect,
+      ...dayList,
+    ];
   }
 
   async buildView(userId: string): Promise<SlackHomeTabDto> {
