+++ app/src/gui/home-tab/views/day-list-item.builder.ts
@@ -1,18 +1,9 @@
 import { Injectable } from "@nestjs/common";
 import dayjs, { Dayjs } from "dayjs";
-import {
-  Actions,
-  Button,
-  Header,
-  Option,
-  OverflowMenu,
-  StaticSelect,
-  ViewBlockBuilder,
-} from "slack-block-builder";
+import { Actions, Button, Header, Option, OverflowMenu, StaticSelect } from "slack-block-builder";
 import Action from "../../../bolt/enums/action.enum";
 import { Office } from "../../../entities/office/office.model";
 import { Presence } from "../../../entities/presence/presence.model";
-import { BlockBuilder } from "../../block-builder.interface";
 
 type DayListItemProps = {
   date: Dayjs;
@@ -21,7 +12,7 @@ type DayListItemProps = {
 };
 
 @Injectable()
-export class DayListItemBuilder implements BlockBuilder<ViewBlockBuilder> {
+export class DayListItem {
   buildBlocks(props: DayListItemProps) {
     const { date } = props;
     const dateString = date.toISOString();
